INT/INT = INT Declares initial values for variable locations. Adds '-1' constant. Quadruples for expressions and assignments almost done.

+++ Entrega 5/Cube.py
@@ -56,7 +56,7 @@ varDictionary['float*float'] = FLOAT
 varDictionary['float*int'] = FLOAT
 
 #Division
-varDictionary['int/int'] = FLOAT
+varDictionary['int/int'] = INT
 varDictionary['int/float'] = FLOAT
 varDictionary['float/int'] = FLOAT
 varDictionary['float/float'] = FLOAT
